requirements.txt
pygame
tello-python
hidapi
easytello